# BREAKING(streams): have `toJson()` and `toText()` accept `ReadableStream<Uint8Array>`