--- a/streams/to_blob.ts
+++ b/streams/to_blob.ts
@@ -3,7 +3,8 @@
 
 /**
  * Converts a {@linkcode ReadableStream} of {@linkcode Uint8Array}s to a
- * {@linkcode Blob}. Works the same as {@linkcode Response.blob}.
+ * {@linkcode Blob}. Works the same as {@linkcode Request.blob} and
+ * {@linkcode Response.blob}.
  *
  * @param stream A `ReadableStream` of `Uint8Array`s to convert into a `Blob`.
  * @returns A `Promise` that resolves to the `Blob`.