--- a/streams/to_json_test.ts
+++ b/streams/to_json_test.ts
@@ -6,18 +6,5 @@ import { toJson } from "./to_json.ts";
 Deno.test("toJson()", async () => {
   const byteStream = ReadableStream.from(["[", "1, 2, 3, 4", "]"])
     .pipeThrough(new TextEncoderStream());
-
   assertEquals(await toJson(byteStream), [1, 2, 3, 4]);
-
-  const stringStream = ReadableStream.from([
-    '{ "a": 2,',
-    ' "b": 3,',
-    ' "c": 4 }',
-  ]);
-
-  assertEquals(await toJson(stringStream), {
-    a: 2,
-    b: 3,
-    c: 4,
-  });
 });