--- a/streams/to_text_test.ts
+++ b/streams/to_text_test.ts
@@ -4,12 +4,7 @@ import { assertEquals } from "@std/assert/assert-equals";
 import { toText } from "./to_text.ts";
 
 Deno.test("toText()", async () => {
-  const byteStream = ReadableStream.from(["hello", " js ", "fans"])
+  const stream = ReadableStream.from(["hello", " js ", "fans"])
     .pipeThrough(new TextEncoderStream());
-
-  assertEquals(await toText(byteStream), "hello js fans");
-
-  const stringStream = ReadableStream.from(["hello", " deno ", "world"]);
-
-  assertEquals(await toText(stringStream), "hello deno world");
+  assertEquals(await toText(stream), "hello js fans");
 });