### What's changed

Previously, [`toJson()`](https://jsr.io/@std/streams/doc/to-json/~/toJson) and [`toText()`](https://jsr.io/@std/streams/doc/to-text/~/toText) accepted `ReadableStream`, without the type of stream being defined. Now, these functions accept `ReadableStream<Uint8Array>`.

### Why this change was made

This change was made to align `toJson()` to [`Request.json()`](https://developer.mozilla.org/en-US/docs/Web/API/Request/json) and [`Response.json()`](https://developer.mozilla.org/en-US/docs/Web/API/Response/json), and `toText()` to [`Request.text()`](https://developer.mozilla.org/en-US/docs/Web/API/Request/text) and [`Response.text()`](https://developer.mozilla.org/en-US/docs/Web/API/Response/text), which seems to be the most common use case for these functions. The change also aligned these two functions with the similarly-oriented [`toBlob()`](https://jsr.io/@std/streams/doc/to-blob/~/toBlob), which did correctly accept `ReadableStream<Uint8Array>`.

### Migration guide

To migrate, ensure that a `ReadableStream` of `Uint8Array`s is being passed to these functions:
```diff
  import { toText } from "@std/streams/to-json";

- const stream = ReadableStream(["Hello, ", "world!"]);
+ const stream = ReadableStream(["Hello, ", "world!"]).pipeThrough(new TextEncoderStream());
  await toText(stream);
```

In most cases, adding `.pipeThrough(new TextEncoderStream())` for existing use cases that use `ReadableStream<string>` is all that's needed to migrate.

### Discussion

Alternatively, all 3 functions could be modified to accept `ReadbleStream<Uint8Array> | ReadableStream<string>`. However, this wasn't done in order to align with the `Request` and `Response` objects and slightly improve performance while still catering to the majority of use cases.