@kt3k 2024-06-19
> This change was made to align toJson() to [Request.json()](https://developer.mozilla.org/en-US/docs/Web/API/Request/json) and [Response.json()](https://developer.mozilla.org/en-US/docs/Web/API/Response/json), and toText() to [Request.text()](https://developer.mozilla.org/en-US/docs/Web/API/Request/text) and [Response.text()](https://developer.mozilla.org/en-US/docs/Web/API/Response/text),

I don't understand why we need to align them to Response methods. This seems unnecessarily reducing the use cases.