@kt3k 2024-06-20
The original inspiration for these APIs are [stream/consumers](https://nodejs.org/api/webstreams.html#webstreams_utility_consumers) APIs of Node.js (see https://github.com/denoland/deno_std/issues/3587 ), and they accept streams of strings