@iuioiua 2024-06-20
Then why don't users use `node:stream/consumers`? It's implemented cleanly in Deno for server-side use cases. For most browser-based use cases, one would likely just need the `Request` and `Response` methods.