@iuioiua 2024-06-20
Yes, I suggest removing them. Frankly, I've never understood the reasons for adding them in the first place 😅