@kt3k 2024-06-20
Maybe let's keep discussing that as RC feedback to `streams`