@iuioiua 2024-07-19
This almost slipped past us before stabilizing `@std/streams`. What would you like to do, @kt3k? I suggest proceeding with this PR, as it seems to support almost all use cases ([GH search](https://github.com/search?q=%2Fstd.*%5C%2Fstreams%2F%20%22toText%22&type=code)).