@kt3k 2024-07-19
I still don't understand why excluding string (and other type) input is desirable. Why do we need to align the behavior to `Response.prototype.toJson`? What's the benefit?