@iuioiua 2024-07-19
I think a better question is why should we support `ReadableStream<string>` when `ReadableStream<Uint8Array>` is the primary use case? These 3 functions were deliberately created to align with `Response` methods (it's in the their descriptions), but extending them to also support `ReadableStream<string>` seems like it was an arbitrary decision. As a Standard Library, we should be consistent and not do more than we need to, across the board.