@kt3k 2024-07-19
> why should we support ReadableStream<string> when ReadableStream<Uint8Array> is the primary use case?

Because they are exposed as public API and we should avoid unnecessary breaking changes as far as possible.

> These 3 functions were deliberately created to align with Response methods (it's in the their descriptions)

It's also explicitly documented that they accept string inputs. So this is intentional design, not a mistake.