@kt3k 2024-07-19
Also there seem no downside of accepting `ReadableStream<string>`. Isn't it just sensible/predictable enhancement to `Response.prototype.json`? We sometimes do such enhancement to Web spec inspired APIs. e.g. `crypto.digest` of `std/crypto` accepts `Iterable` and `AsyncIterable` data sources in addition to ArrayBuffer(View)s