@kt3k 2024-07-23
> as most are using ReadableStream<Uint8Array>.

I disagree with this observation. There are several std APIs that return `ReadableStream<string>`:
- `CsvStringifyStream`
- `JsonStringifyStream`
- `TextDelimiterStream`
- `TextLineStream`

Also this change causes perf penalty when the user has `ReadableStream<string>`. With this change the user once needs to encode it into `ReadableStream<Uint8Array>`, then decode it to text again using `toText`.