@iuioiua 2024-07-23
Ok. I'll submit a superseding PR that clarifies these functions and unblocks `streams@1.0.0`.